Refactor input manager and rendering surfaces for SDL integration
- Updated `SimpleDisplayInputManager` to use a class-level vtable for input handling, improving performance and organization.
- Replaced `BackendWindow` and `WindowRenderSurface` with `SDLWindowRenderSurface` for better compatibility with SDL.
- Simplified the `Window` class to utilize the new SDL surface, removing legacy code.
- Removed unused render surface classes and consolidated rendering logic.
- Enhanced event handling in `RenderingManager` to improve viewport management and debugging output.
- Ensured proper cleanup of resources in the new SDL-based architecture.

+++ cpp/termin/camera/orbit_camera_controller.cpp
@@ -233,16 +233,16 @@ void OrbitCameraController::on_mouse_button(void* event) {
     ViewportState& state = _get_viewport_state(vp_key);
 
     // Middle mouse = orbit
-    if (e->button == MouseButton::Middle) {
-        state.orbit_active = (e->action == Action::Press);
+    if (e->button == static_cast<int>(MouseButton::Middle)) {
+        state.orbit_active = (e->action == static_cast<int>(Action::Press));
     }
     // Right mouse = pan
-    else if (e->button == MouseButton::Right) {
-        state.pan_active = (e->action == Action::Press);
+    else if (e->button == static_cast<int>(MouseButton::Right)) {
+        state.pan_active = (e->action == static_cast<int>(Action::Press));
     }
 
     // Reset last position on release
-    if (e->action == Action::Release) {
+    if (e->action == static_cast<int>(Action::Release)) {
         state.has_last = false;
     }
 }
